functions/requirements.txt
firebase_functions~=0.1.0
coinbase-advanced-py==1.8.1
firebase-functions
firebase-admin
python-dotenv
sendgrid